ci/github-script: extract common withRateLimit.js
This can and should be re-used across different scripts.
